vInteger, vReal type
new types to accomodate for verilog style "real" "integer" and typeless
parameters. convert when crossing boundaries.

verilog "math mode" flag. effectively maths operations depend on scope
they were entered in.

passes tests, but needs cleanup. possible bugs around default values and
NOT_INPUT.

+++ apps/c_param.cc
@@ -47,10 +47,10 @@ void parse(CS& cmd, PARAM_LIST* pl)
     PARAM_INSTANCE par;
     switch(type){
     case 2:
-      par = PARAMETER<int>();
+      par = PARAMETER<Integer>();
       break;
     default:
-      par = PARAMETER<double>();
+      par = PARAMETER<Float>();
       break;
     }
